# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Luna Negra · API para desarrolladores",
    "version": "1.0.0",
    "description": "Contrato público de Luna Negra para desarrolladores de juegos.\n\nAutenticación: los tokens de acceso (entitlement) e invitación (sala) van en el header `Authorization: Bearer <token>`. Son JWT firmados con ES256: podés validarlos **offline** con la clave pública de `/.well-known/jwks.json` (recomendado) o contra los endpoints `/verify` de abajo.\n\nErrores: forma estándar `{ \"error\": { \"code\", \"message\" } }` con el status HTTP correspondiente."
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": { "host": { "default": "luna-negra-three.vercel.app" } }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token emitido por Luna Negra (entitlement o invite)."
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key del proveedor (`ln_sk_…`), creada en el panel /provider."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "MISSING_TOKEN" },
              "message": { "type": "string" }
            },
            "required": ["code", "message"]
          }
        }
      },
...
      "CreateBetRequest": {
        "type": "object",
        "required": ["gameId", "participants", "stakeSats"],
        "properties": {
          "gameId": { "type": "string" },
          "participants": {
            "type": "array",
            "items": { "type": "string", "description": "npub" },
            "minItems": 2
          },
          "stakeSats": {
            "type": "integer",
            "description": "Monto por jugador (sats). Debe estar entre minStakeSats y maxStakeSats; fuera de rango → STAKE_OUT_OF_RANGE."
          },
          "victoryCondition": { "type": "string" },
          "roomId": { "type": "string", "description": "Sala multijugador del juego (correlación, opcional)" },
          "metadata": { "type": "object", "description": "Metadata libre; se persiste y vuelve en GET y en cada webhook" }
        }
      },
      "BetEconomics": {
        "type": "object",
        "description": "Economía de la apuesta (sats).",
        "properties": {
          "stakeSats": { "type": "integer", "description": "Stake por jugador" },
          "potTargetSats": { "type": "integer", "description": "Pozo total cuando esté completo (stake × participantes)" },
          "feePct": { "type": "integer", "description": "Comisión de Luna Negra (%)" },
          "feeBps": { "type": "integer", "description": "Comisión en basis points (feePct × 100)" },
          "feeSats": { "type": "integer", "description": "Comisión absoluta" },
          "netPayoutSats": { "type": "integer", "description": "Pago neto a repartir entre ganadores (pozo − comisión)" }
        }
      },
      "CreateBetResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BetEconomics" },
          {
            "type": "object",
            "properties": {
              "betId": { "type": "string" },
              "contractEventId": { "type": ["string", "null"], "description": "Evento Nostr del contrato firmado" },
              "depositDeadline": { "type": "string", "format": "date-time" },
              "roomId": { "type": ["string", "null"] },
              "metadata": { "type": ["object", "null"] }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/bets": {
      "post": {
        "summary": "Crear una apuesta (escrow)",
        "description": "El game server crea una apuesta; Luna Negra custodia el pozo y publica el contrato firmado en Nostr. Autenticación con API key del proveedor.",
        "tags": ["Apuestas"],
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Clave única para reintentos seguros: un reintento con la misma key devuelve la respuesta original sin crear otra apuesta.",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBetRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Apuesta creada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateBetResponse" }
              }
            }
          },
          "401": {
            "description": "API key inválida",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "El juego no es de tu proveedor",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
...
  },
...
}